[#13] Reorder by Dependencies and minor refactors from review
