+++ build.gradle.kts
@@ -17,6 +17,7 @@ plugins {
     id(Plugins.ANDROID_LIBRARY) version Versions.GRADLE apply (false)
     id(Plugins.DETEKT) version Versions.DETEKT
     id(Plugins.KOVER) version Versions.KOVER
+    id(Plugins.KSP) version Versions.KSP
     kotlin(Plugins.ANDROID) version Versions.KOTLIN apply (false)
     kotlin(Plugins.MULTIPLATFORM) version Versions.KOTLIN apply (false)
 }
